refactored: Clean implementation of parseFirstStep
parseFirstStep is supposed to correlate the induction variable
in (the lhs of) the proposition with its instances (the
lhs of) the equations in the proof. However, parseStep just
accumulated all instances, without any sanity check. Use 'match`
to only return a result if the variable is instantiated the
same way on all occurrences.